GOT IT TOO WORK
FINALLY got it to do what i wanted it to do, it does it in a round about way and is way more overcomplicated than it probably needed to be, but its editable in engine and should still run fine with any length of information for the bounties. im calling it good and will be showing its progress in class monday.